requirements.txt
django
django-extensions
django-debug-toolbar
djangorestframework
djangorestframework-simplejwt
dj-rest-auth
djoser
pillow